pyproject.toml
[project]
name = "fullstack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi", 
    "uvicorn",
    "psycopg2-binary",
]
